Allow the Wrapper to be backed up and restored.
This is handy if you want the compression dictionary to persist
across application restarts. This will avoid the "warm up" period where
the library is constructing the compression dictionary, during which
time the data is not compressed.

It also means that the wrapped data structures can be persisted to disk.
Otherwise, the decompression dictionary would be lost when the
application exits, making any saved data worthless.

Note that the Backup() operation will fail if a compression dictionary
has not yet been generated; it would not make sense to create a backup
without one.